add x64 build for windows and fix all problems. fix maclib for x64 use searchctrl from wxwidgets 2.8 for SearchBox change comboboxes to radioboxes in selection prefs (much easier to select static/dynamic). fix flac for x64 on windows make path relative automatically on library path setup, if wxMusik is used as portable app (.musik directory exists next to the wxMusik.exe) and the directory is on the same drive as wxMusik.exe of course finish transparent dialog setting work.

+++ trunk/wxMusik/MUSIKEngine/3rd-Party/MAC/MACLib/APEDecompress.cpp
@@ -367,9 +367,9 @@ int CAPEDecompress::SeekToFrame(int nFrameIndex)
 /*****************************************************************************************
 Get information from the decompressor
 *****************************************************************************************/
-int CAPEDecompress::GetInfo(APE_DECOMPRESS_FIELDS Field, int nParam1, int nParam2)
+intptr_t CAPEDecompress::GetInfo(APE_DECOMPRESS_FIELDS Field, intptr_t nParam1,intptr_t nParam2)
 {
-    int nRetVal = 0;
+    intptr_t nRetVal = 0;
     BOOL bHandled = TRUE;
 
     switch (Field)
@@ -452,7 +452,7 @@ int CAPEDecompress::GetInfo(APE_DECOMPRESS_FIELDS Field, int nParam1, int nParam
             }
             else
             {
-                WAVEFORMATEX wfeFormat; GetInfo(APE_INFO_WAVEFORMATEX, (int) &wfeFormat, 0);
+                WAVEFORMATEX wfeFormat; GetInfo(APE_INFO_WAVEFORMATEX, (intptr_t) &wfeFormat, 0);
                 WAVE_HEADER WAVHeader; FillWaveHeader(&WAVHeader, 
                     (m_nFinishBlock - m_nStartBlock) * GetInfo(APE_INFO_BLOCK_ALIGN), 
                     &wfeFormat,    0);
